backend/requirements.txt
accelerate
aiohttp
aiosignal
attrs
blinker
certifi
charset-normalizer
click
colorama
coloredlogs
contourpy
cycler
datasets
dill
et-xmlfile
filelock
Flask
Flask-Cors
fonttools
frozenlist
fsspec
google-generativeai>=0.8.0
huggingface-hub>=0.23.0
humanfriendly
idna
intel-openmp
itsdangerous
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib
mkl
mpmath
multidict
multiprocess
networkx
numpy
openai>=1.0.0
openpyxl
optimum
packaging
pandas
pillow
plotly>=5.18.0
protobuf
psutil
pyarrow
pyarrow-hotfix
pyparsing
pyreadline3
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentencepiece
six
sympy
tbb
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers>=4.36.0
typing_extensions
tzdata
urllib3
Werkzeug
xxhash
yarl
